sourcedb/requirements.txt
# fastapi
sqlalchemy
sqlmodel
python-dotenv
pyodbc
# psycopg2
# PyMySQL
# oracledb

aiofiles==22.1.0
celery==5.2.7
fastapi==0.90.0
flower==1.2.0
Jinja2==3.1.2
pytest==7.2.1
redis==4.5.0
requests==2.28.2
uvicorn==0.20.0

loguru